patch a possible floating error
In certain rare cases, the previous epsilon caused the following error
in card_pos: ValueError: min() arg is an empty sequence.

This is just a patch. A more sensible correction would use a closed
formula to calculate the intersection between the box and the quadratic
Bézier curve. Not implemented yet.