add preliminary support for invisible boxes
An entity with the special name "phantom" suffixed by at least one digit is not displayed. It still can be referenced, notably as a constraint position, but also in an association, which may lead to pending legs.
A better way to do this would use a "-" prefix before the name of an entity or an association. If an association is invisible, all its legs are invisible too.